International Journal on Data Science and Technology
Abstract: The process of evaluating data using analytical and logical reasoning to examine each component of the data
provided. This form of analysis is just one of the many steps that must be completed when conducting a research
experiment. Data from various sources is gathered, reviewed, and then analyzed to form some sort of finding or conclusion.
There are a variety of specific data analysis method, some of which include data mining, text analytics, business
intelligence, and data visualizations. Data analysis is defined as a process of cleaning, transforming, and modeling data to
discover useful information for business decision-making. The purpose of Data Analysis is to extract useful information
from data and taking the decision based upon the data analysis. whenever we take any decision in our day-to-day life is by
thinking about what happened last time or what will happen by choosing that particular decision. This is nothing but
analyzing our past or future and making decisions based on it. For that, we gather memories of our past or dreams of our
future. So that is nothing but data analysis. Now same thing analyst does for business purposes, is called Data Analysis.
This research article based on data analysis, it’s types, process, methods, techniques & tools.