2019 International Conference on Information Science and Communications Technologies (ICISCT), At Tashkent, Uzbekistan
Abstract — Data visualization is a general term that describes any
effort to help people understand the significance of data by placing
it in a visual context. Patterns, trends and correlations that might
go undetected in text-based data can be exposed and recognized
easier with data visualization software. Data visualization is the
presentation of quantitative information in a graphical form. In
other words, data visualizations turn large and small data-sets into
visuals that are easier for the human brain to understand and
process. Data visualizations are surprisingly common in our
everyday life, but they often appear in the form of well-known
charts and graphs. It can be used to discover unknown facts and
trends. Good data visualizations are created when communication,
data science, and design collide. Data visualizations done right
offer key insights into complicated data-sets in ways that are
meaningful and intuitive. In this article , we would like to discuss
about data visualization, importance of data visualization, data
visualization tools etc.